- introduced "initializeRadio" method in BasePhyLayer which creates and initializes the radio - replaced public constructor of radio with static factory method to enable sub classes to change the amount of radio states - made public methods of Radio virtual
- fixed parameter of DrawAmounts constructor from int to double
- updated Signals documentation
- fixed a warning regarding wrong initialization order of parameters in ConcatConstMappings Constructor

+++ base/modules/BaseBattery.h
@@ -35,7 +35,7 @@ protected:
 	double value;
 
 public:
-	DrawAmount(int type = CURRENT, int value = 0):
+	DrawAmount(int type = CURRENT, double value = 0):
 		type(type),
 		value(value)
 	{}
